DBZ-5749 Convert Redis store into standard store module
