+++ debezium-bom/pom.xml
@@ -40,6 +40,9 @@
         <version.groovy>3.0.7</version.groovy>
         <version.graalvm.js>20.0.0</version.graalvm.js>
 
+        <!-- Storages -->
+        <version.jedis>4.1.1</version.jedis>
+
         <!-- Testing -->
         <version.junit>4.13.1</version.junit>
         <version.fest>1.4</version.fest>
@@ -423,6 +426,13 @@
                 <scope>provided</scope>
             </dependency>
 
+            <!--Storages -->
+            <dependency>
+                <groupId>redis.clients</groupId>
+                <artifactId>jedis</artifactId>
+                <version>${version.jedis}</version>
+            </dependency>
+
             <!-- Testing utilities -->
             <dependency>
                 <groupId>org.testcontainers</groupId>
@@ -534,10 +544,10 @@
                 <artifactId>debezium-scripting</artifactId>
                 <version>${project.version}</version>
             </dependency>
-	    <dependency>
+            <dependency>
                 <groupId>io.debezium</groupId>
                 <artifactId>debezium-scripting-languages</artifactId>
-		<type>pom</type>
+                <type>pom</type>
                 <version>${project.version}</version>
             </dependency>
             <dependency>
@@ -618,6 +628,11 @@
             <dependency>
                 <groupId>io.debezium</groupId>
                 <artifactId>debezium-quarkus-outbox-deployment</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.debezium</groupId>
+                <artifactId>debezium-storage-redis</artifactId>
                 <version>${project.version}</version>
             </dependency>
