+++ debezium-storage/pom.xml
@@ -11,9 +11,15 @@
     <name>Debezium Storage Module</name>
     <packaging>pom</packaging>
 
+    <properties>
+        <!-- Assembly configuration -->
+        <assembly.descriptor>storage-distribution</assembly.descriptor>
+    </properties>
+
     <modules>
         <module>debezium-storage-kafka</module>
         <module>debezium-storage-file</module>
+        <module>debezium-storage-redis</module>
     </modules>
 </project>
